requirements.txt
flask
Flask-SQLAlchemy
gunicorn
python-dotenv
mysqlclient
cloudinary
Pillow
cryptography